Fix: auth callback now returns to last page
I moved all metadata into a service and made all pages wait for app
instance, selection and do anything unless one is selected (could
think on it a bit more, but works for now)
