+++ client/demo/src/app/auth/auth-callback/auth-callback.component.html
@@ -1,3 +1,1 @@
-<p>
-  auth-callback works!
-</p>
+<p>please wait ...</p>
