+++ client/demo/src/app/auth/auth-callback/auth-callback.component.ts
@@ -15,7 +15,6 @@ export class AuthCallbackComponent implements OnInit {
 	ngOnInit() {
 		this.authService.completeAuthentication().then(() => {
 			const returnUrl = sessionStorage.getItem('returnUrl');
-			console.log(returnUrl);
 			this.router.navigateByUrl(returnUrl);
 			sessionStorage.removeItem('returnUrl');
 		});
