--- client/src/lib/common/services/metadata.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { MetadataService } from './metadata.service';
-
-describe('MetadataService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [MetadataService]
-    });
-  });
-
-  it('should be created', inject([MetadataService], (service: MetadataService) => {
-    expect(service).toBeTruthy();
-  }));
-});
