+++ client/src/lib/list/list-container/list-container.component.html
@@ -1,1 +1,12 @@
-<dscribe-list [entity]="entity"></dscribe-list>
+<div class="list-wrapper">
+	<div class="list-container">
+		<nav class="secondary-nav mat-elevation-z4">
+			<a class="secondary-nav-item" *ngFor="let entity of entities" routerLink="/entity/{{entity.name}}"
+				 routerLinkActive="active">
+				{{entity.pluralTitle || entity.name}}
+			</a>
+		</nav>
+
+		<dscribe-list [entity]="entity"></dscribe-list>
+	</div>
+</div>
