+++ client/src/lib/navigation/navigation.component.ts
@@ -1,6 +1,4 @@
 import {Component, OnInit} from '@angular/core';
-import {MetadataService} from '../common/services/metadata.service';
-import {EntityMetadata} from '../metadata/entity-metadata';
 import {Router} from '@angular/router';
 import {HttpClient} from '@angular/common/http';
 import {AppInstanceInfoModel} from '../common/models/app-instance-info-model';
@@ -9,22 +7,17 @@ import {DscribeService} from '../dscribe.service';
 @Component({
 	selector: 'dscribe-navigation',
 	templateUrl: './navigation.component.html',
-	styleUrls: ['./navigation.component.css'],
+	styleUrls: ['./navigation.component.scss'],
 })
 export class NavigationComponent implements OnInit {
-	entities: EntityMetadata[];
-	mainUrls = ['entity', 'administration'];
-	sideNavOpen = true;
 	appInstances: AppInstanceInfoModel[] = [];
 	selectedAppInstance: AppInstanceInfoModel;
 
-	constructor(private metadata: MetadataService, private router: Router, private httpClient: HttpClient,
-							private metaData: MetadataService, private config: DscribeService) {
+	constructor(private router: Router, private httpClient: HttpClient, private config: DscribeService) {
+
 	}
 
 	ngOnInit() {
-		this.navigate(this.mainUrls[0]);
-
 		this.httpClient.post<AppInstanceInfoModel[]>(this.config.url('api/AppManagement/getAppInstancesInfo'), null)
 			.subscribe(apps => {
 				this.appInstances = apps;
@@ -32,21 +25,8 @@ export class NavigationComponent implements OnInit {
 			});
 	}
 
-	navigate(url: string) {
-		this.router.navigateByUrl(url);
-	}
-
 	appInstanceSelected(appInstance: AppInstanceInfoModel) {
 		this.config.appInstance = appInstance;
 		this.selectedAppInstance = appInstance;
-		this.metadata.clearMetadata();
-		this.metadata.getMetadata().getAllTypes()
-			.subscribe(entities => {
-				this.entities = entities;
-				if (this.entities && this.entities.length) {
-					this.mainUrls[0] = 'entity/' + this.entities[0].name;
-					this.router.navigateByUrl(this.mainUrls[0]);
-				}
-			});
 	}
 }
