+++ client/src/lib/themes/_core.scss
@@ -72,10 +72,3 @@
 	font: 500 20px/32px IranSans, Metropolis, sans-serif;
 	margin-top: 0;
 }
-
-
-.container main dscribe-list-container {
-	display: block;
-	margin-left: 240px;
-}
-
